+++ src/modules/eightface/EightFace.cpp
@@ -322,7 +322,8 @@ struct EightFaceModule : Module {
 										for (int i = 0; i < presetCount; i++) {
 											slotCvModeShuffle.push_back(i);
 										}
-										std::random_shuffle(std::begin(slotCvModeShuffle), std::end(slotCvModeShuffle));
+										std::mt19937 rng(random::u32());
+										std::shuffle(std::begin(slotCvModeShuffle), std::end(slotCvModeShuffle), rng);
 									}
 									int p = std::min(std::max(0, slotCvModeShuffle.back()), presetCount - 1);
 									slotCvModeShuffle.pop_back();