+++ src/modules/intermix/Intermix.cpp
@@ -228,10 +228,7 @@ struct IntermixModule : Module, IntermixBase<PORTS> {
 			if (resetTrigger.process(inputs[INPUT_RESET].getVoltage())) {
 				resetTimer.reset();
 				switch (sceneMode) {
-					case SCENE_CV_MODE::TRIG_FWD:
-					case SCENE_CV_MODE::TRIG_RANDOM:
-					case SCENE_CV_MODE::TRIG_RANDOM_WALK:
-					case SCENE_CV_MODE::TRIG_RANDOM_WO_REPEAT: {
+					case SCENE_CV_MODE::TRIG_FWD: {
 						sceneSet(0);
 						break;
 					}
@@ -255,12 +252,16 @@ struct IntermixModule : Module, IntermixBase<PORTS> {
 						for (int i = 0; i < sceneCount; i++) {
 							sceneCvModeShuffle.push_back(i);
 						}
-						std::random_shuffle(std::begin(sceneCvModeShuffle), std::end(sceneCvModeShuffle));
+						std::mt19937 rng(random::u32());
+						std::shuffle(std::begin(sceneCvModeShuffle), std::end(sceneCvModeShuffle), rng);
 						int s = std::min(std::max(0, sceneCvModeShuffle.back()), sceneCount - 1);
 						sceneCvModeShuffle.pop_back();
 						sceneSet(s);
 						break;
 					}
+					case SCENE_CV_MODE::TRIG_RANDOM:
+					case SCENE_CV_MODE::TRIG_RANDOM_WALK:
+					case SCENE_CV_MODE::TRIG_RANDOM_WO_REPEAT:
 					default: {
 						break;
 					}
@@ -353,7 +354,8 @@ struct IntermixModule : Module, IntermixBase<PORTS> {
 							for (int i = 0; i < sceneCount; i++) {
 								sceneCvModeShuffle.push_back(i);
 							}
-							std::random_shuffle(std::begin(sceneCvModeShuffle), std::end(sceneCvModeShuffle));
+							std::mt19937 rng(random::u32());
+							std::shuffle(std::begin(sceneCvModeShuffle), std::end(sceneCvModeShuffle), rng);
 						}
 						int s = std::min(std::max(0, sceneCvModeShuffle.back()), sceneCount - 1);
 						sceneCvModeShuffle.pop_back();