+++ src/utils/string.hpp
@@ -1,0 +1,24 @@
+#pragma once
+#include <string>
+
+namespace StoermelderPackOne {
+namespace strings {
+
+const std::string WHITESPACE = " \n\r\t\f\v";
+
+inline std::string ltrim(const std::string& s) {
+	size_t start = s.find_first_not_of(WHITESPACE);
+	return (start == std::string::npos) ? "" : s.substr(start);
+}
+
+inline std::string rtrim(const std::string& s) {
+	size_t end = s.find_last_not_of(WHITESPACE);
+	return (end == std::string::npos) ? "" : s.substr(0, end + 1);
+}
+
+inline std::string trim(const std::string& s) {
+	return rtrim(ltrim(s));
+}
+
+} // namespace strings
+} // namespace StoermelderPackOne